pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "commandor-ai"
version = "0.1.0"
description = "Agentic CLI - Autonomous coding assistant with multi-provider AI support"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
  { name = "Ravin D", email = "ravin.d3107@outlook.com" }
]

keywords = [
  "terminal", "ai", "shell", "command-line", "natural-language",
  "gemini", "claude", "gpt", "assistant", "automation", 
  "cli", "productivity", "agent", "opencode", "codex"
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Shells",
  "Topic :: System :: System Shells",
  "Topic :: Terminals",
  "Topic :: Utilities",
  "Topic :: Artificial Intelligence"
]

dependencies = [
  "google-genai>=1.0.0",
  "anthropic>=0.40.0",
  "openai>=1.0.0",
  "httpx>=0.27.0",
  "rich>=13.0.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "ruff>=0.1.0",
  "mypy>=1.0.0",
]
windows = [
  "pyreadline3>=3.4.1"
]

[project.urls]
Homepage = "https://github.com/ravin-d-27/Commandor"
BugReports = "https://github.com/ravin-d-27/Commandor/issues"
Source = "https://github.com/ravin-d-27/Commandor"
Documentation = "https://github.com/ravin-d-27/Commandor#readme"

[project.scripts]
commandor = "commandor.__main__:main"
